requirements.txt
attrs==24.3.0
barcode==1.0.4
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
cryptography-fernet==0.1.0
idna==3.10
imageio==2.36.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
Levenshtein==0.26.1
llvmlite==0.43.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
PyMatting==1.1.13
python-barcode==0.15.1
python-Levenshtein==0.26.1
qrcode==8.0
RapidFuzz==3.12.1
referencing==0.36.1
rembg==2.0.61
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.15.1
tifffile==2025.1.10
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
